Error : in Admin. problem, calling controller twice when delete
-  삭제 부분에서 controller가 두번씩 호출되는 문제
(시점 - admin repo에서 일부를 optional로 변경한 후. 원인 - entitiy manager와 관련 있을 것으로 추정, 동일한 형식의 Book에서는 현상이 발생하지 않음. 때문에 차이점인, security와의 연관이나, id를 문자형으로 지정한 것과 관련이 있지 않을까 싶음.)
- 두가지 방편, 하나는 deleteByUid를 만들어서 @query를 사용하거나, controller에서 null체크를 하도록하는 것. 우선, 후자를 적용함.
- 이후, 화면에 반영 안되는 이슈가 생겨서 list에서 무조건 한번 리로드하도록 변경

+++ src/main/java/com/evaluation/controller/AdminController.java
@@ -6,6 +6,7 @@ import com.evaluation.service.AdminService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,6 +19,7 @@ import lombok.extern.slf4j.Slf4j;
 @Controller
 @RequestMapping("/admin/**")
 @Slf4j
+@Transactional
 public class AdminController {
 
     @Setter(onMethod_ = { @Autowired })
@@ -68,10 +70,12 @@ public class AdminController {
     public String remove(String uid, RedirectAttributes rttr) {
         log.info("delete " + uid);
 
-        adminService.remove(uid);
-
+        if (uid != null) {
+            adminService.remove(uid);
+        }
         rttr.addFlashAttribute("msg", "remove");
         return "redirect:/admin/list";
+
     }
 
     @GetMapping("/list")
